refactor quick save
