+++ codeforlife/__init__.py
@@ -6,3 +6,5 @@ DATA_DIR = BASE_DIR.joinpath("data")
 
 
 from .version import __version__
+
+from . import user
