+++ codeforlife/user/admin.py
@@ -1,29 +1,39 @@
 from django.contrib import admin
 from django.contrib.auth.admin import UserAdmin
 
-from .models import CFL_User
+from .models import User
 
 
 class CustomUserAdmin(UserAdmin):
-    model = CFL_User
+    model = User
 
-    list_display = ('username', 'email', 'is_active',
-                    'is_staff', 'is_superuser', 'last_login',)
-    list_filter = ('is_active', 'is_staff', 'is_superuser')
+    list_display = (
+        "username",
+        "email",
+        "is_active",
+        "is_staff",
+        "is_superuser",
+        "last_login",
+    )
+
+    list_filter = ("is_active", "is_staff", "is_superuser")
+
     fieldsets = (
-        (None, {'fields': ('username', 'email', 'password')}),
-        ('Permissions', {'fields': ('is_staff', 'is_active',
-         'is_superuser', 'groups', 'user_permissions')}),
-        ('Dates', {'fields': ('last_login', 'date_joined')})
+        (None, {"fields": ("username", "email", "password")}),
+        ("Permissions", {"fields": ("is_staff", "is_active", "is_superuser", "groups", "user_permissions")}),
+        ("Dates", {"fields": ("last_login", "date_joined")}),
     )
+
     add_fieldsets = (
-        (None, {
-            'classes': ('wide',),
-            'fields': ('username', 'email', 'password1', 'password2', 'is_staff', 'is_active')}
-         ),
+        (
+            None,
+            {"classes": ("wide",), "fields": ("username", "email", "password1", "password2", "is_staff", "is_active")},
+        ),
     )
-    search_fields = ('email',)
-    ordering = ('email',)
+
+    search_fields = ("email",)
+
+    ordering = ("email",)
 
 
-admin.site.register(CFL_User, CustomUserAdmin)
+admin.site.register(User, CustomUserAdmin)
