+++ codeforlife/user/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class UserConfig(AppConfig):
-    name = "user"
+    name = "codeforlife.user"
