+++ codeforlife/user/models/__init__.py
@@ -1,0 +1,9 @@
+from .classroom import Class
+
+# from .other import *
+from .school import School
+from .session import UserSession
+from .student import Student
+from .teacher_invitation import SchoolTeacherInvitation
+from .teacher import Teacher
+from .user import User
