+++ codeforlife/user/models/session.py
@@ -1,0 +1,17 @@
+from django.db import models
+from django.utils import timezone
+
+from .user import User
+from .school import School
+from .classroom import Class
+
+
+class UserSession(models.Model):
+    cfl_user = models.ForeignKey(User, on_delete=models.CASCADE)
+    login_time = models.DateTimeField(default=timezone.now)
+    school = models.ForeignKey(School, null=True, on_delete=models.SET_NULL)
+    class_field = models.ForeignKey(Class, null=True, on_delete=models.SET_NULL)
+    login_type = models.CharField(max_length=100, null=True)  # for student login
+
+    def __str__(self):
+        return f"{self.cfl_user} login: {self.login_time} type: {self.login_type}"
