+++ codeforlife/user/tests/test_user.py
@@ -1,31 +1,28 @@
 import django
 import pytest
 
+from django.contrib.auth.hashers import make_password
+
+
 django.setup()
+
+from ..models import User
 
 
 @pytest.mark.django_db
 def test_user_import():
     """Basic test to ensure importing the package does not raise an error."""
 
-    from user.models import CFL_User
+    user_kwargs = {
+        "username": "test teacher",
+        "first_name": "Albert",
+        "last_name": "Einstein",
+        "email": "alberteinstein@codeforlife.com",
+        "password": make_password("Password1"),
+    }
 
-
-    from django.contrib.auth.hashers import make_password
-
-    user = CFL_User.objects.create(
-        username="test teacher",
-        first_name="Albert",
-        last_name="Einstein",
-        email="alberteinstein@codeforlife.com",
-        password=make_password("Password1"),
-    )
-
-    user.save()
-
-    created_user = CFL_User.objects.get(username="test teacher")
-    assert created_user.username == "test teacher"
-    assert created_user.first_name == "Albert"
-    assert created_user.last_name == "Einstein"
-    assert created_user.email == "alberteinstein@codeforlife.com"
-
+    user = User.objects.create_user(**user_kwargs)
+    assert user.username == user_kwargs["username"]
+    assert user.first_name == user_kwargs["first_name"]
+    assert user.last_name == user_kwargs["last_name"]
+    assert user.email == user_kwargs["email"]
