+++ manage.py
@@ -1,19 +1,8 @@
-"""
-Django settings for mysite project.
-
-Generated by 'django-admin startproject' using Django 3.2.12.
-
-For more information on this file, see
-https://docs.djangoproject.com/en/3.2/topics/settings/
-
-For the full list of settings and their values, see
-https://docs.djangoproject.com/en/3.2/ref/settings/
-"""
-
 from pathlib import Path
 
+
 # Build paths inside the project like this: BASE_DIR / 'subdir'.
-BASE_DIR = Path(__file__).resolve().parent.parent
+BASE_DIR = Path(__file__).resolve().parent
 
 
 # Quick-start development settings - unsuitable for production
@@ -38,7 +27,7 @@ INSTALLED_APPS = [
     "django.contrib.messages",
     "django.contrib.staticfiles",
     "django.contrib.sites",
-    "user.apps.UserConfig",
+    "codeforlife.user.apps.UserConfig",
     "django_extensions",
 ]
 
@@ -52,7 +41,7 @@ MIDDLEWARE = [
     "django.middleware.clickjacking.XFrameOptionsMiddleware",
 ]
 
-ROOT_URLCONF = "user.urls"
+ROOT_URLCONF = "codeforlife.user.urls"
 
 TEMPLATES = [
     {
@@ -70,7 +59,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = "user.wsgi.application"
+WSGI_APPLICATION = "codeforlife.user.wsgi.application"
 
 
 # Database
@@ -102,7 +91,7 @@ AUTH_PASSWORD_VALIDATORS = [
     },
 ]
 
-AUTH_USER_MODEL = "user.CFL_User"
+AUTH_USER_MODEL = "user.User"
 
 # Internationalization
 # https://docs.djangoproject.com/en/3.2/topics/i18n/
@@ -127,3 +116,14 @@ STATIC_URL = "/static/"
 # https://docs.djangoproject.com/en/3.2/ref/settings/#default-auto-field
 
 DEFAULT_AUTO_FIELD = "django.db.models.BigAutoField"
+
+
+if __name__ == "__main__":
+
+    from django.core.management import execute_from_command_line
+    import sys
+    import os
+
+    os.environ.setdefault("DJANGO_SETTINGS_MODULE", Path(__file__).stem)
+
+    execute_from_command_line(sys.argv)
